+++ src/Assets/css/Style.css
@@ -41,3 +41,12 @@ padding-bottom: 17px;
 .imgcls{
     filter: drop-shadow(2px 4px 6px black);
 }
+p.errors {
+  color: darkred;
+}
+.fmcls{
+margin-bottom: 40px;
+}
+button.btn.btn-primary {
+  margin-top: 27px;
+}
